requirements.txt
Django==3.1.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt
dj-database-url==0.5.0
psycopg2==2.9.3
python-decouple==3.6
Pillow==9.1.0
gunicorn

coverage
django-fake-model==0.1.4
django-filter==21.1
django-phonenumber-field[phonenumberslite]
django-storages[azure]==1.12.3
django-polymorphic==3.1.0
freezegun==1.2.1
ipdb==0.13.9
model-bakery==1.5.0
parameterized==0.8.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-subtests==0.7.0
pytest==7.1.1

xhtml2pdf==0.2.7
